Shallow Redd 2024 Winter-run 2024-10-31 JC
Attached is the 6600 remeasurements with dewatering flows revisited. ID Number 15 has been downgraded to 6,000, and 17 is being downgraded to 6400. Number 15 is near a heavy amount of Fall-run spawning on the kayak ramp, so some of the tailspill has been impacted. Number 17 is near the shoreline, so I am trying to use Mark Gard's model in conjunction with the historical redds to fill in the gaps. Marks model predicted it dewatering at 6600 (which did not happen) and it is only a foot off the shoreline, so it is hard to tell when it will happen. I updated my emergence calculations tonight and we are still on track for 3 redds emerging on Monday November 4th.